requirements.txt
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
geopy==2.4.1
